+++ Assets/Scripts/Gameplay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c1255ec0a3488344a4e8a51aa6a82ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
