+++ Assets/Scripts/Gameplay/GameManager.cs.meta
