+++ Assets/Scripts/PlayerRelated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ccafd81ff8fb87746baf9b61fde3422e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
